authentication and authorization for contacts routes

+++ controllers/contacts.js
@@ -1,8 +1,10 @@
 const contactsService = require("../services/contacts");
 const Contact = require("../services/schemas/contact");
+const validators = require("../utils/validators");
+
 const getContacts = async (req, res, next) => {
   try {
-    res.json(await contactsService.get());
+    res.json(await contactsService.getByOwner(req.user.id));
   } catch (e) {
     console.log(e.message);
     next(e);
@@ -13,10 +15,13 @@ const getContactById = async (req, res, next) => {
   try {
     const contact = await contactsService.getById(req.params.id);
     if (contact) {
-      res.json(contact);
-      return;
+      if (contact.owner === req.user.id) {
+        res.json(contact);
+        return;
+      }
+      return res.status(401).json({ message: "Not Authorized" });
     }
-    res.status(404).json({ message: "Not found" });
+    return res.status(404).json({ message: "Not found" });
   } catch (e) {
     console.log(e.message);
     next(e);
@@ -26,20 +31,27 @@ const getContactById = async (req, res, next) => {
 const addContact = async (req, res, next) => {
   try {
     const contact = req.body;
+    contact.owner = req.user.id;
     const addResult = await contactsService.add(contact);
-    res.status(addResult._id ? 201 : 400).json(addResult.message);
+    return res.status(addResult._id ? 201 : 400).json({ addResult });
   } catch (e) {
     console.log(e.message);
-    next(e);
+    return next(e);
   }
 };
 
 const deleteContact = async (req, res, next) => {
   try {
-    const removed = await contactsService.remove(req.params.id);
-    if (removed._id) {
-      res.json({ message: "contact deleted" });
-      return;
+    const foundContact = await contactsService.getById(req.params.id);
+    if (foundContact) {
+      if (`${foundContact.owner}` === req.user.id) {
+        const removed = await contactsService.remove(req.params.id);
+        if (removed._id) {
+          return res.json({ message: "contact deleted" });
+        }
+        return res.status(400).json({ removed });
+      }
+      return res.status(401).json({ message: "Not Authorized" });
     }
     res.status(404).json({ message: "Not found" });
   } catch (e) {
@@ -66,18 +78,20 @@ const prepareContact = (contact) => {
 const updateContact = async (req, res, next) => {
   try {
     const contact = req.body;
+    contact.owner = req.user.id;
     const contactInvalidity = isContactInvalid(contact);
-    if (contactInvalidity) {
+    if (!validators.validateContact(contact)) {
       res.status(400).json(contactInvalidity);
     }
-
-    const updateResult = await contactsService.update(req.params.id, prepareContact(contact));
-    if (updateResult._id) {
-      res.status(200).json(updateResult);
-      return;
+    const foundContact = await contactsService.getById(req.params.id);
+    if (`${foundContact.owner}` === req.user.id) {
+      const updateResult = await contactsService.update(req.params.id, prepareContact(contact));
+      if (updateResult._id) {
+        return res.status(200).json(updateResult);
+      }
+      return res.status(404).json(updateResult);
     }
-    res.status(404).json(updateResult);
-    return;
+    return res.status(401).json({ message: "Not Authorized" });
   } catch (e) {
     console.log(e.message);
     next(e);
@@ -91,10 +105,14 @@ const setFavorite = async (req, res, next) => {
       res.status(400).json({ message: "missing field favorite" });
       return;
     }
-    const patchResult = await contactsService.update(req.params.id, {
-      favorite: isFavorite,
-    });
-    res.status(patchResult._id === undefined ? 404 : 200).json(patchResult);
+    const foundContact = await contactsService.getById(req.params.id);
+    if (`${foundContact.owner}` === req.user.id) {
+      const patchResult = await contactsService.update(req.params.id, {
+        favorite: isFavorite,
+      });
+      return res.status(patchResult._id === undefined ? 404 : 200).json(patchResult);
+    }
+    return res.status(401).json({ message: "Not Authorized" });
   } catch (e) {
     console.log(e.message);
     next(e);
